Re-organized the tests under \contract.
Re-organized the tests under \contract so that we can use them to cover the scenario where latest Net Native tools + last released stable packages are used.

Ref #1048.
Fix #1121.
