+++ src/System.Private.ServiceModel/tests/Scenarios/Contract/Data/Contract.Data.Tests.csproj
@@ -17,7 +17,7 @@
     <ProjectGuid>{4361D851-B56C-44CD-B7BD-D9EFD32F94AD}</ProjectGuid>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="DataContractTests.cs" />
+    <Compile Include=".\*.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="project.json" />
@@ -32,7 +32,7 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
       <Targets>Build;DebugSymbolsProjectOutputGroup</Targets>
     </ProjectReference>
-    <ProjectReference Include='$(WcfScenarioTestCommonProj)'>
+    <ProjectReference Include="$(WcfScenarioTestCommonProj)">
       <Project>{9098B41C-9C2E-4FC6-B7D8-FC3411736A22}</Project>
       <Name>ScenarioTests.Common</Name>
     </ProjectReference>
