+++ src/System.Private.ServiceModel/tests/Scenarios/Contract/Fault/Contract.Fault.Tests.csproj
@@ -17,7 +17,7 @@
     <ProjectGuid>{0324E738-B6EF-43FB-8321-DC49640D2615}</ProjectGuid>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="FaultExceptionTests.cs" />
+    <Compile Include=".\*.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="project.json" />
@@ -32,7 +32,7 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
       <Targets>Build;DebugSymbolsProjectOutputGroup</Targets>
     </ProjectReference>
-    <ProjectReference Include='$(WcfScenarioTestCommonProj)'>
+    <ProjectReference Include="$(WcfScenarioTestCommonProj)">
       <Project>{9098B41C-9C2E-4FC6-B7D8-FC3411736A22}</Project>
       <Name>ScenarioTests.Common</Name>
     </ProjectReference>
