+++ src/System.Private.ServiceModel/tests/Scenarios/Contract/Message/Contract.Message.Tests.csproj
@@ -17,9 +17,7 @@
     <ProjectGuid>{A835C2BC-DFD0-4ED6-A240-59FBEE7F0157}</ProjectGuid>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="MessageContractCommon.cs" />
-    <Compile Include="MessageContractTests.cs" />
-    <Compile Include="MessageTests.cs" />
+    <Compile Include=".\*.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="project.json" />
@@ -34,7 +32,7 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
       <Targets>Build;DebugSymbolsProjectOutputGroup</Targets>
     </ProjectReference>
-    <ProjectReference Include='$(WcfScenarioTestCommonProj)'>
+    <ProjectReference Include="$(WcfScenarioTestCommonProj)">
       <Project>{9098B41C-9C2E-4FC6-B7D8-FC3411736A22}</Project>
       <Name>ScenarioTests.Common</Name>
     </ProjectReference>
