+++ src/System.Private.ServiceModel/tests/Scenarios/Contract/Service/Contract.Service.Tests.csproj
@@ -17,8 +17,7 @@
     <ProjectGuid>{8773AC5C-5B4D-4745-BECF-01415CB10F1C}</ProjectGuid>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="ServiceContractTests.cs" />
-    <Compile Include="ServiceKnownTypeTests.cs" />
+    <Compile Include=".\*.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="project.json" />
@@ -33,7 +32,7 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
       <Targets>Build;DebugSymbolsProjectOutputGroup</Targets>
     </ProjectReference>
-    <ProjectReference Include='$(WcfScenarioTestCommonProj)'>
+    <ProjectReference Include="$(WcfScenarioTestCommonProj)">
       <Project>{9098B41C-9C2E-4FC6-B7D8-FC3411736A22}</Project>
       <Name>ScenarioTests.Common</Name>
     </ProjectReference>
