+++ src/System.Private.ServiceModel/tests/Scenarios/Contract/XmlSerializer/Contract.XmlSerializer.Tests.csproj
@@ -17,7 +17,7 @@
     <ProjectGuid>{DA0F8A1F-30A4-4AD4-9C5E-F6743D7ACEFE}</ProjectGuid>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="XmlSerializerFormatTest.cs" />
+    <Compile Include=".\*.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="project.json" />
@@ -32,7 +32,7 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
       <Targets>Build;DebugSymbolsProjectOutputGroup</Targets>
     </ProjectReference>
-    <ProjectReference Include='$(WcfScenarioTestCommonProj)'>
+    <ProjectReference Include="$(WcfScenarioTestCommonProj)">
       <Project>{9098B41C-9C2E-4FC6-B7D8-FC3411736A22}</Project>
       <Name>ScenarioTests.Common</Name>
     </ProjectReference>
